refactor(coprocessor): share solc ETXTBSY retry helper
