+++ coprocessor/fhevm-engine/host-listener/build.rs
@@ -1,41 +1,11 @@
 use foundry_compilers::{
-    error::SolcError,
-    multi::MultiCompiler,
-    solc::{Solc, SolcCompiler},
-    Project, ProjectPathsConfig,
+    multi::MultiCompiler, solc::SolcCompiler, Project, ProjectPathsConfig,
 };
 use semver::Version;
-use std::{env, fs, path::Path, process::Command, thread, time::Duration};
+use std::{env, fs, path::Path, process::Command};
 
-const SOLC_INSTALL_RETRY_ATTEMPTS: usize = 5;
-
-fn is_etxtbsy(err: &SolcError) -> bool {
-    matches!(err, SolcError::Io(io_err) if io_err.source().kind() == std::io::ErrorKind::ExecutableFileBusy || io_err.source().raw_os_error() == Some(26))
-}
-
-fn find_or_install_solc(version: &Version) -> Solc {
-    for attempt in 1..=SOLC_INSTALL_RETRY_ATTEMPTS {
-        match Solc::find_or_install(version) {
-            Ok(solc) => return solc,
-            Err(err)
-                if is_etxtbsy(&err)
-                    && attempt < SOLC_INSTALL_RETRY_ATTEMPTS =>
-            {
-                let backoff = Duration::from_millis((attempt as u64) * 200);
-                println!(
-                    "cargo:warning=solc install race (ETXTBSY), retrying attempt {attempt}/{SOLC_INSTALL_RETRY_ATTEMPTS} after {:?}",
-                    backoff
-                );
-                thread::sleep(backoff);
-            }
-            Err(err) => {
-                panic!("Failed to find or install solc {version}: {err:?}")
-            }
-        }
-    }
-
-    unreachable!("solc install retry loop exited unexpectedly")
-}
+#[path = "../build-utils/solc_retry.rs"]
+mod solc_retry;
 
 fn build_contracts() {
     println!(
@@ -113,7 +83,7 @@ fn main() {
             .unwrap();
     // Use a specific version due to an issue with libc and libstdc++ in the
     // rust Docker image we use to run it.
-    let solc = find_or_install_solc(&Version::new(0, 8, 28));
+    let solc = solc_retry::find_or_install_solc(&Version::new(0, 8, 28));
     let project = Project::builder()
         .paths(paths)
         .build(
